chatbot/pyproject.toml
[project]
name = "chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "conversation-states",
    "faker==37.1.0",
    "hypercorn==0.17.3",
    "langchain==0.3.23",
    "langchain-community==0.3.21",
    "langchain-core==0.3.54",
    "langchain-openai==0.3.14",
    "langchain-text-splitters==0.3.8",
    "langgraph==0.4.3",
    "langgraph-api==0.2.21",
    "langgraph-cli==0.2.10",
    "langgraph-prebuilt==0.1.8",
    "langgraph-sdk==0.1.66",
    "langsmith==0.3.29",
    "python-dotenv==1.1.0",
    "python-telegram-bot==22.0",
    "quart==0.20.0",
    "requests==2.32.3",
    "tavily-python==0.5.4",
    "trustcall==0.0.36",
    "watchdog[watchmedo]==6.0.0",
    "wikipedia==1.4.0",
]

[tool.setuptools]
packages = []